svip_api/requirements.txt
backcall==0.1.0
biothings-client==0.2.1
CacheControl==0.12.5
cachecontrol-django==1.0.0
certifi==2018.10.15
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.2
Django==2.2.4
django-cors-headers==2.4.0
django-db-cascade-2==0.2.2
django-debug-toolbar==2.0
django-filter==2.0.0
django-simple-history==2.7.3
djangorestframework==3.9.0
djangorestframework-simplejwt==3.2.3
drf-nested-routers==0.91
drf-writable-nested==0.5.4
django-redis==4.12.1
redis==3.5.3
redgreenunittest==0.1.1
drf-yasg==1.11.0
flex==6.13.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
ipython==7.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.0
msgpack==0.6.1
mygene==3.1.0
nose==1.3.7
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.8
psycopg2==2.8.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.6.4
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-futures==1.0.0
rfc3987==1.3.8
ruamel.yaml==0.15.77
six==1.11.0
sqlparse==0.3.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
wcwidth==0.1.7
